pyproject.toml
[tool.poetry]
name = "open-retrieval"
version = "0.1.0"
description = ""
authors = ["Jack <i@jackpan.me>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
fastapi = "^0.110.0"
uvicorn = { extras = ["standard"], version = "^0.28.0" }
pydantic = "^2.6.4"
python-ulid = "^2.2.0"
llama-index-core = "^0.10.20.post2"
llama-index-embeddings-voyageai = "^0.1.3"
llama-index-readers-file = "^0.1.11"
llama-index-vector-stores-pinecone = "^0.1.4"
pinecone-client = "^3.1.0"
tree-sitter = "^0.21.1"
tree-sitter-languages = "^1.10.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
